+++ .gitignore
@@ -10,3 +10,4 @@ client/dist
 .MultiSelectExample.vue
 .HomeTest.vue
 .GeneDistribution1.vue
+DevTemp
