+++ client/app/components/viz/SummaryPieChart.vue
